requirements.txt
allure-pytest-bdd==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
parse==1.19.0
parse-type==0.6.0
phonenumbers==8.13.11
playwright==1.33.0
pluggy==1.0.0
pyee==9.0.4
pytest==7.3.1
pytest-base-url==2.0.0
pytest-bdd==6.1.1
pytest-playwright==0.3.3
python-slugify==8.0.1
requests==2.30.0
six==1.16.0
sttable==0.0.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
